requirements.txt
# Core dependencies
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Machine learning
scikit-learn>=1.3.0
sentence-transformers>=2.2.0

# Retry and rate limiting
tenacity>=8.2.0
ratelimit>=2.2.1

# Logging
loguru>=0.7.0

# Experiment tracking
wandb>=0.15.0

# Text processing
python-Levenshtein>=0.21.0
fuzzywuzzy>=0.18.0

# CLI
click>=8.1.0
tqdm>=4.65.0

# Testing (dev)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1

